moving dataclasses out of upload
