+++ src/ingest_validation_tools/error_report.py
@@ -1,13 +1,122 @@
-from typing import List, Optional, Union
+from collections import defaultdict
+from dataclasses import dataclass, field, fields
+from datetime import datetime
+from pathlib import Path
+from typing import DefaultDict, Dict, List, Optional, Union
 
 from yaml import Dumper, dump
 
 from ingest_validation_tools.message_munger import munge, recursive_munge
-from ingest_validation_tools.upload import ErrorDict, InfoDict
 
 # Force dump not to use alias syntax.
 # https://stackoverflow.com/questions/13518819/avoid-references-in-pyyaml
 Dumper.ignore_aliases = lambda *args: True
+
+
+class ErrorDictException(Exception):
+    def __init__(self, error: str):
+        super().__init__(error)
+        self.error = error
+
+
+@dataclass
+class InfoDict:
+    time: datetime
+    git: str
+    dir: str
+    tsvs: Dict[str, Dict[str, str]]
+
+    def as_dict(self):
+        return {
+            "Time": self.time,
+            "Git version": self.git,
+            "Directory": self.dir,
+            # "Directory schema version": self.dir_schema,
+            "TSVs": self.tsvs,
+        }
+
+
+@dataclass
+class ErrorDict:
+    """
+    Has fields for each major validation type, which can be accessed directly or
+    compiled using self.as_dict().
+    """
+
+    preflight: List[str] = field(default_factory=list)
+    directory: DefaultDict[str, List[str]] = field(default_factory=lambda: defaultdict(list))
+    upload_metadata: DefaultDict[str, List[str]] = field(default_factory=lambda: defaultdict(list))
+    metadata_validation_local: DefaultDict[str, List[str]] = field(
+        default_factory=lambda: defaultdict(list)
+    )
+    metadata_validation_api: DefaultDict[str, List] = field(
+        default_factory=lambda: defaultdict(list)
+    )
+    metadata_url_errors: DefaultDict[str, List] = field(default_factory=lambda: defaultdict(list))
+    reference: DefaultDict[str, Dict] = field(default_factory=lambda: defaultdict(dict))
+    plugin: Dict[str, List[str]] = field(default_factory=dict)
+    plugin_skip: Optional[str] = None
+
+    def __bool__(self):
+        """
+        Return true if any field has errors.
+        """
+        return bool(self.as_dict())
+
+    @property
+    def field_map(self):
+        """
+        Single source of truth for top-level error dict key names.
+        """
+        return {
+            "preflight": "Preflight Errors",
+            "directory": "Directory Errors",
+            "upload_metadata": "Antibodies/Contributors Errors",
+            "metadata_validation_local": "Local Validation Errors",
+            "metadata_validation_api": "API Validation Errors",
+            "metadata_url_errors": "URL Check Errors",
+            "reference": "Reference Errors",
+            "plugin": "Data File Errors",
+            "plugin_skip": "Fatal Errors",
+        }
+
+    def tsv_only_errors_by_path(self, path: str, local_allowed=False):
+        """
+        For use in front-end single TSV validation.
+        """
+        errors = {}
+        for metadata_field in [
+            "metadata_validation_local",
+            "metadata_url_errors",
+            "metadata_validation_api",
+        ]:
+            if metadata_field == "metadata_validation_local" and not local_allowed:
+                continue
+            for key, value in getattr(self, metadata_field).items():
+                if Path(key) == Path(path):
+                    errors[self.field_map.get(metadata_field)] = self.sort_val(value)
+                    break
+        return errors
+
+    def as_dict(self):
+        """
+        Compiles all fields with errors into a dict.
+        """
+        errors = {}
+        for error_field in fields(self):
+            value = getattr(self, error_field.name)
+            if value:
+                value = self.sort_val(value)
+                errors[self.field_map.get(error_field.name)] = value
+        return errors
+
+    def sort_val(self, value):
+        """
+        Recursively sort all dicts by keys for consistency of testing and output.
+        """
+        if type(value) in [dict, defaultdict]:
+            value = {k: self.sort_val(v) for k, v in sorted(value.items())}
+        return value
 
 
 class ErrorReport:
